feat: 리뷰 작성 예약 시간 비교

+++ src/main/java/com/example/tripKo/domain/place/dto/response/info/HomeResponse.java
@@ -27,7 +27,7 @@ public class HomeResponse {
     private String summary;
     private String image;
     private String address;
-    private Float averageRating;
+    private double averageRating;
     private Boolean liked;
   }
